Rewrite build system

+++ build.sh
@@ -5,14 +5,14 @@ love=$2
 if [ $1 == '--bare' ]; then
     love=lib4.love
 
-    cd $2
+    cd $2/
     echo "Creating $love..."
     
     rm -f $dir/$love
     
     cd lib4/
     zip -9 -r $dir/$love main.lua conf.lua
-    cd ..
+    cd ../
     
     for f in $(find lib4 -type f); do
         [ $f != 'lib4/main.lua' ] && zip -9 -r $dir/$love $f
@@ -32,21 +32,22 @@ if [ $1 == '--bare' ]; then
         mv lib/$lib/init.lua lib/$lib/$lib.lua
     done
 
-    cd $dir
+    cd $dir/
     
     echo "Done."
-else
+elif [ -f 'lib4.love' ]; then
     echo "Creating $love..."
+
+    rm -f $dir/$love
+
+    tmp=$(mktemp -d)
+    unzip -d $tmp lib4.love
     
     rm -f $dir/$love
-    
-    cd lib4/
-    zip -9 -r $dir/$love main.lua conf.lua
+
+    cd $tmp
+    zip -9 -r $dir/$love *
     cd $dir
-    
-    for f in $(find lib4 -type f); do
-        [ $f != 'lib4/main.lua' ] && zip -9 -r $dir/$love $f
-    done
     
     cd $1/
     for f in $(find . -type f -regex '.*\.lua'); do
@@ -56,7 +57,25 @@ else
     for f in src assets LICENSE README.md; do
         [[ -e ./$f ]] && zip -9 -r $dir/$love $f
     done
-    cd $dir
+    cd $dir/
+    
+    echo "Done."
+else
+    echo "Creating $love..."
+    
+    rm -f $dir/$love
+
+    if [[ -e $dir/lib4/lib4 ]]; then
+        cd lib4
+    fi
+    
+    cd lib4/
+    zip -9 -r $dir/$love main.lua conf.lua
+    cd $dir/
+    
+    for f in $(find lib4 -type f); do
+        [ $f != 'lib4/main.lua' ] && zip -9 -r $dir/$love $f
+    done
     
     for lib in cpml iqm love3d; do
         for f in $(find lib/$lib -type f); do
@@ -71,6 +90,20 @@ else
         done
         mv lib/$lib/init.lua lib/$lib/$lib.lua
     done
+
+    if [[ -e $dir/lib4/lib4 ]]; then
+        cd $dir
+    fi
+    
+    cd $1/
+    for f in $(find . -type f -regex '.*\.lua'); do
+        [[ $f != *.git* ]] && zip -9 -r $dir/$love $f
+    done
+    
+    for f in src assets LICENSE README.md; do
+        [[ -e ./$f ]] && zip -9 -r $dir/$love $f
+    done
+    cd $dir/
     
     echo "Done."
 fi
